[BUGFIX:BP:10] Handle custom java command options for server module as well
Related: #174
