+++ Classes/Service/Tika/AbstractService.php
@@ -17,6 +17,7 @@ namespace ApacheSolrForTypo3\Tika\Service\Tika;
 use Psr\Log\LoggerAwareInterface;
 use Psr\Log\LoggerAwareTrait;
 use Psr\Log\LogLevel;
+use TYPO3\CMS\Core\Utility\CommandUtility;
 
 /**
  * Abstract Tika service implementing shared methods
@@ -26,6 +27,8 @@ use Psr\Log\LogLevel;
  */
 abstract class AbstractService implements ServiceInterface, LoggerAwareInterface
 {
+    protected const JAVA_COMMAND_OPTIONS_REGEX = '/-D(?P<property>[\w.]+)=(?P<value>"[^"]+"|\'[^\']+\'|[^\\s\'"]+)/';
+
     use LoggerAwareTrait;
 
     /**
@@ -79,5 +82,38 @@ abstract class AbstractService implements ServiceInterface, LoggerAwareInterface
      */
     public function getSupportedMimeTypes() {
         return [];
+    }
+
+    /**
+     * Parse additional Java command options.
+     *
+     * Reads the configuration value `javaCommandOptions` and tries to parse it to a
+     * safe argument string. For safety reasons, only the following variants are
+     * allowed (multiple separated by space):
+     *
+     * -Dfoo=bar
+     * -Dfoo='hello world'
+     * -Dfoo="hello world"
+     *
+     * @return string Parsed additional Java command options
+     */
+    protected function getAdditionalCommandOptions(): string
+    {
+        $commandOptions = trim((string)($this->configuration['javaCommandOptions'] ?? ''));
+
+        // Early return if no additional command options are configured
+        // or configuration does not match required pattern (only -D parameter is supported)
+        if ('' === $commandOptions || !preg_match_all(self::JAVA_COMMAND_OPTIONS_REGEX, $commandOptions, $matches)) {
+            return '';
+        }
+
+        // Combine matched command options with escaped argument value
+        $commandOptionsString = '';
+        foreach (array_combine($matches['property'], $matches['value']) as $property => $unescapedValue) {
+            $escapedValue = CommandUtility::escapeShellArgument(trim($unescapedValue, '"\''));
+            $commandOptionsString .= sprintf(' -D%s=%s', $property, $escapedValue);
+        }
+
+        return $commandOptionsString;
     }
 }
