+++ Classes/Service/Tika/ServerService.php
@@ -94,6 +94,8 @@ class ServerService extends AbstractService
      */
     protected function getProcess($arguments = ''): Process
     {
+        $arguments = trim($this->getAdditionalCommandOptions() . ' ' . $arguments);
+
         return GeneralUtility::makeInstance(Process::class, CommandUtility::getCommand('java'), $arguments);
     }
